@@ -14,7 +14,8 @@
     "batchelor": "^2.0.2",
     "cheerio": "^1.0.0-rc.2",
     "dateformat": "^2.0.0",
-    "googleapis": "^21.2.0",
+    "google-auth-library": "^0.10.0",
+    "googleapis": "^21.3.0",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "redux": "^3.6.0",
@@ -29,12 +30,10 @@
     "@types/react-dom": "^15.5.0",
     "@types/redux": "^3.6.0",
     "aws-sdk": "^2.104.0",
-    "google-auth-library": "^0.10.0",
-    "googleapis": "^21.2.0",
     "json-loader": "^0.5.7",
-    "serverless-webpack": "^2.2.2",
+    "serverless-webpack": "^2.2.3",
     "ts-loader": "^2.3.4",
     "typescript": "^2.4.2",
-    "webpack": "^3.5.5"
+    "webpack": "^3.6.0"
   }
 }
